Add stackoverflow question description update checker

+++ scrapper/src/main/java/edu/java/webClients/stackOverflow/StackOverflowClient.java
@@ -10,9 +10,9 @@ import org.springframework.web.service.annotation.HttpExchange;
 @HttpExchange
 public interface StackOverflowClient {
     @GetExchange(url = "/questions/{id}?site=stackoverflow&filter=withbody")
-    StackOverflowResponseBody<StackOverflowQuestionBody> fetchQuestionById(@PathVariable int id);
+    StackOverflowResponseBody<StackOverflowQuestionBody> fetchQuestionById(@PathVariable long id);
 
     @GetExchange(url = "/questions/{id}/answers?site=stackoverflow&filter=withbody")
-    StackOverflowResponseBody<StackOverflowAnswerBody> fetchAnswersByQuestionId(@PathVariable int id);
+    StackOverflowResponseBody<StackOverflowAnswerBody> fetchAnswersByQuestionId(@PathVariable long id);
 
 }
